Create speedtest_streamlit.py
This is a internet speedtest in py using streamlit and send the data to your database and while it runs, it keeps get the data for speed test every 1 minutes. To run a streamlit code,  use command streamlit run .\your_filename.py don't forget to install streamlit, 
pip install streamlit and other necessities.